refactor: 패키지 이름, 구조 변경, 불필요한 클래스 제거
