+++ app/src/main/java/com/stormers/storm/canvas/fragment/CanvasTextFragment.kt
@@ -1,8 +1,8 @@
-package com.stormers.storm.round.fragment
+package com.stormers.storm.canvas.fragment
 
 import android.os.Bundle
 import android.view.View
-import com.stormers.storm.base.BaseCanvasFragment
+import com.stormers.storm.canvas.base.BaseCanvasFragment
 import kotlinx.android.synthetic.main.fragment_round_canvas.*
 
 class CanvasTextFragment : BaseCanvasFragment(TEXT_MODE) {
